create question widget
There is an issue with the build method which is reordering the alternatives every time the widget is rebuilt (e.g. resizing the screen). I have no knowledge as of yet on how to handle this conundrum without removing `const` from the constructor. Therefore, I am leaving it as is.